Support wave_vpi

+++ src/verilua/lua_vpi.cpp
@@ -3,6 +3,7 @@
 #include "vpi_access.h"
 #include "vpi_callback.h"
 #include "vpi_user.h"
+#include <cstdint>
 
 lua_State *L;
 bool verilua_is_init = false;
@@ -11,7 +12,7 @@ sol::protected_function sim_event;
 sol::protected_function main_step; 
 
 IDPool edge_cb_idpool(50);
-std::unordered_map<int, vpiHandle> edge_cb_hdl_map;
+std::unordered_map<uint64_t, vpiHandle> edge_cb_hdl_map;
 std::unordered_map<std::string, vpiHandle> handle_cache;
 std::unordered_map<vpiHandle, VpiPrivilege_t> handle_cache_rev;
 bool enable_vpi_learn = false;
